examples/requirements.txt
minio==5.0.10
PyJWT==1.7.1
requests==2.23.0
urllib3==1.25.9
xmltodict==0.12.0
